ダミーイメージ生成関数を共通関数にくくりだし
